Labs 5+6 Updated redability and fixed ex5
