+++ src/lab/java/basics/lab_1_6/ex1and2/BankWorker.java
@@ -1,6 +1,6 @@
-package lab.java.basics.lab_1_6;
+package lab.java.basics.lab_1_6.ex1and2;
 
-public class Client extends Human {
+public class BankWorker extends Human implements BankWorkerInter {
     public String bankName;
 
     @Override
@@ -11,10 +11,18 @@ public class Client extends Human {
         System.out.println("Bank Name: " + bankName);
     }
 
-    public Client(String name, String surname, String bankName) {
+    public BankWorker(String name, String surname, String bankName) {
         this.name = name;
         this.surname = surname;
         this.bankName = bankName;
+
+        greet();
+    }
+
+    @Override
+    public void performDuties() {
+        System.out.println("Performing bank worker duties...");
+
     }
 
 }
