+++ src/lab/java/basics/lab_1_6/ex3/Car.java
@@ -1,4 +1,4 @@
-package lab.java.basics.lab_1_6;
+package lab.java.basics.lab_1_6.ex3;
 
 //1.3
