+++ src/lab/java/basics/lab_1_6/ex4/MainClass.java
@@ -1,4 +1,4 @@
-package lab.java.basics.lab_1_6;
+package lab.java.basics.lab_1_6.ex4;
 //1.4
 
 public class MainClass {
