+++ src/lab/java/basics/lab_1_6/ex5/DescendantAge.java
@@ -1,13 +1,13 @@
-package lab.java.basics.lab_1_6;
+package lab.java.basics.lab_1_6.ex5;
 
 //1.5
 
 class DescendantAge extends MainAge {
 
     @Override
-    public String getStuff() {
+    public void getStuff() {
         System.out.println("OVERRIDEN: Enter your username");
         String age = newscanner.nextLine();
-        return age;
+        System.out.println("Username: "+ age);
     }
 }
