+++ src/lab/java/basics/lab_1_6/ex5/MainAge.java
@@ -1,4 +1,4 @@
-package lab.java.basics.lab_1_6;
+package lab.java.basics.lab_1_6.ex5;
 
 //1.5
 
@@ -7,14 +7,13 @@ import java.util.Scanner;
 class MainAge {
     Scanner newscanner = new Scanner(System.in);
 
-    public String getStuff() {
+    public void getStuff() {
         System.out.println("Enter your age");
         String age = newscanner.nextLine();
-        return age;
     }
 
     public static void main(String[] args) {
         DescendantAge childOne = new DescendantAge();
-        System.out.println(childOne.getStuff());
+        childOne.getStuff();
     }
 }
